+++ devicemodel/include/acpi.h
@@ -59,5 +59,8 @@ void	dsdt_unindent(int levels);
 void	sci_init(struct vmctx *ctx);
 void	pm_write_dsdt(struct vmctx *ctx, int ncpu);
 void	pm_backto_wakeup(struct vmctx *ctx);
+void	inject_power_button_event(struct vmctx *ctx);
+void	power_button_init(struct vmctx *ctx);
+void	power_button_deinit(struct vmctx *ctx);
 
 #endif /* _ACPI_H_ */
